exercises/04-multi-agent-ai/requirements.txt
numpy>=1.24.0,<2.0.0
pytest>=7.4.0,<8.0.0
black>=23.7.0,<24.0.0
pyyaml>=6.0,<7.0.0
# Add specific dependencies after reviewing grand_solution.ipynb
